Added Stocks and Crypto database tables, changed id fields to Long and generated getters and setters for all tables.

+++ src/main/java/edu/carroll/cs389/jpa/model/Login.java
@@ -9,22 +9,22 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
 @Entity
-@Table(name="login")
+@Table(name = "login")
 public class Login {
     private static final long serialVersionUID = 1L;
 
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
 
     @Column(name = "username", nullable = false, unique = true)
     private String username;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
